pyproject.toml
[project]
name = "countdown-tray"
version = "2.0.0"
description = "Create a system tray icon that counts down"
authors = [
    {name = "Ethan Dawes", email = "ethand.python@gmail.com"},
]
dependencies = ["pystray>=0.19.5", "pillow>=11.0.0", "cronsim>=2.6"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/funblaster22/countdown-tray"
Issues = "https://github.com/funblaster22/countdown-tray/issues"

[project.scripts]
countdown-tray = "countdown_tray:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
